Moved app download methods to AppRouter
