+++ src/routes/AppRouter.tsx
@@ -2,6 +2,7 @@ import { AmplifyAuthenticator } from "@aws-amplify/ui-react";
 import React, { FC, useState } from "react";
 import { Route } from "react-router-dom";
 import styled from "styled-components";
+import { useReactPWAInstall } from "react-pwa-install";
 import Header from "../components/Header";
 import MenuBar from "../components/MenuBar";
 import MyAdverts from "../components/MyAdverts";
@@ -97,6 +98,15 @@ const AppRouter: FC = () => {
   const [modalOpen, setModalOpen] = useState(false);
   const [alreadyAQRCode, setAlreadyAQRCode] = useState(false);
   const [qrCamera, setQrCamera] = useState({ delay: 500, result: "" });
+  const { pwaInstall, supported, isInstalled } = useReactPWAInstall();
+  const handleClick = () => {
+    pwaInstall({
+      title: "Installera Haffa",
+      description: "Haffa kommer hamna på din hemskärm som en app.",
+    })
+      .then()
+      .catch(() => alert("Ladda gärna ner den nästa gång du använder Haffa."));
+  };
 
   return (
     <AmplifyAuthenticator>
@@ -114,6 +124,11 @@ const AppRouter: FC = () => {
         </SignInContent>
       </SignInWrapper>
       <AppContainer>
+        {supported() && !isInstalled() && (
+          <button type="button" onClick={handleClick}>
+            Lägg Haffa på hemskärmen
+          </button>
+        )}
         <Header isInDetail={false} />
         <Route exact path="/" component={StartScreen} />
         <Route exact path="/onboarding" component={Onboarding} />
